requirements.txt
numpy<2.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pydantic==2.5.0
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
